Conflicts:
	database/src/main/java/cm/aptoide/pt/database/accessors/AccessorFactory.java
	database/src/main/java/cm/aptoide/pt/database/realm/Download.java
	utils/src/main/java/cm/aptoide/pt/utils/FileUtils.java
	v8engine/src/main/java/cm/aptoide/pt/v8engine/fragment/implementations/AppViewFragment.java
	v8engine/src/main/java/cm/aptoide/pt/v8engine/fragment/implementations/DownloadsFragment.java
	v8engine/src/main/java/cm/aptoide/pt/v8engine/install/provider/DownloadInstallationProvider.java
	v8engine/src/main/java/cm/aptoide/pt/v8engine/view/recycler/displayable/implementations/grid/UpdateDisplayable.java
	v8engine/src/main/java/cm/aptoide/pt/v8engine/view/recycler/displayable/implementations/grid/UpdatesHeaderDisplayable.java
	v8engine/src/main/java/cm/aptoide/pt/v8engine/view/recycler/widget/implementations/appView/AppViewInstallWidget.java
	v8engine/src/main/java/cm/aptoide/pt/v8engine/view/recycler/widget/implementations/grid/RollbackWidget.java
	v8engine/src/main/java/cm/aptoide/pt/v8engine/view/recycler/widget/implementations/grid/UpdateWidget.java
	v8engine/src/main/java/cm/aptoide/pt/v8engine/view/recycler/widget/implementations/grid/UpdatesHeaderWidget.java

+++ utils/src/main/java/cm/aptoide/pt/utils/FileUtils.java
@@ -5,15 +5,15 @@
 
 package cm.aptoide.pt.utils;
 
+import android.graphics.Bitmap;
 import android.text.TextUtils;
-
+import cm.aptoide.pt.logger.Logger;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
-
-import cm.aptoide.pt.logger.Logger;
 
 /**
  * Created by trinkes on 5/18/16.
@@ -35,6 +35,10 @@ public class FileUtils {
 	 * @param fileName   Name of the file to be copied
 	 */
 	public static void copyFile(String inputPath, String outputPath, String fileName) {
+		if (!fileExists(inputPath)) {
+			throw new RuntimeException("Input file doesn't exists");
+		}
+
 		File file = new File(inputPath + fileName);
 		if (!file.renameTo(new File(outputPath + fileName))) {
 			cloneFile(inputPath, outputPath, fileName);
@@ -125,5 +129,33 @@ public class FileUtils {
 			}
 		}
 		return result;
+	}
+
+	public static boolean saveBitmapToFile(File dir, String fileName, Bitmap bm,
+			Bitmap.CompressFormat format, int quality) {
+
+		File imageFile = new File(dir, fileName);
+
+		FileOutputStream fos = null;
+		try {
+			dir.mkdirs();
+			fos = new FileOutputStream(imageFile);
+
+			bm.compress(format, quality, fos);
+
+			fos.close();
+
+			return true;
+		} catch (IOException e) {
+			Logger.e(TAG, e.getMessage());
+			if (fos != null) {
+				try {
+					fos.close();
+				} catch (IOException e1) {
+					e1.printStackTrace();
+				}
+			}
+		}
+		return false;
 	}
 }
